requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
pandas>=2.1.0
openpyxl>=3.1.2
python-docx>=1.1.0
docxtpl>=0.17.0
openai>=1.12.0
anthropic>=0.39.0
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.27.0
google-genai>=0.1.0
python-multipart>=0.0.9
pydantic>=2.6.0
pdfplumber>=0.10.0
